# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Film API Documentation",
        "description": "API documentation for Film management",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Main API Server"
        }
    ],
    "paths": {
        "/api/besuche": {
            "get": {
                "tags": [
                    "Besuche"
                ],
                "summary": "Liste aller Kinobesuche",
                "operationId": "af02d529811403ca04634b6bdb8355f1",
                "responses": {
                    "200": {
                        "description": "Liste von Kinobesuchen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Besuch"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Besuche"
                ],
                "summary": "Erstelle einen neuen Kinobesuch",
                "operationId": "6435fe5032c0ddf03bbc2b187bd9ee20",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BesuchInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kinobesuch erfolgreich erstellt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Besuch"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validierungsfehler",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Die angegebenen Daten sind ungültig."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/besuche/{id}": {
            "get": {
                "tags": [
                    "Besuche"
                ],
                "summary": "Zeige einen Kinobesuch",
                "operationId": "b1e10a2cb2d3cea28c07bd91397676eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Kinobesuchs",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kinobesuch",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Besuch"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kinobesuch nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Besuch nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Besuche"
                ],
                "summary": "Aktualisiere einen Kinobesuch",
                "operationId": "74f32b0c2a100e7c28ac919d6e72b908",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Kinobesuchs",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BesuchInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kinobesuch aktualisiert",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Besuch"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kinobesuch nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Besuch nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validierungsfehler",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Die angegebenen Daten sind ungültig."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Besuche"
                ],
                "summary": "Lösche einen Kinobesuch",
                "operationId": "bb8ed520449bd1f1a18599687299f997",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Kinobesuchs",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Kinobesuch gelöscht"
                    },
                    "404": {
                        "description": "Kinobesuch nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Besuch nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Besuch": {
                "required": [
                    "id",
                    "anfang",
                    "ende",
                    "reihe",
                    "platz",
                    "untertitel",
                    "film_id",
                    "typ_id",
                    "sprache_id",
                    "saal_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "anfang": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T18:30:00+00:00"
                    },
                    "ende": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T21:00:00+00:00"
                    },
                    "reihe": {
                        "type": "string",
                        "example": "A"
                    },
                    "platz": {
                        "type": "integer",
                        "example": 12
                    },
                    "untertitel": {
                        "type": "boolean",
                        "example": true
                    },
                    "snackzuschlag_chf": {
                        "type": "number",
                        "format": "float",
                        "example": 3.5,
                        "nullable": true
                    },
                    "film_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "typ_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "sprache_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "saal_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "BesuchInput": {
                "required": [
                    "anfang",
                    "ende",
                    "reihe",
                    "platz",
                    "untertitel",
                    "film_id",
                    "typ_id",
                    "sprache_id",
                    "saal_id"
                ],
                "properties": {
                    "anfang": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T18:30:00+00:00"
                    },
                    "ende": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T21:00:00+00:00"
                    },
                    "reihe": {
                        "type": "string",
                        "example": "A"
                    },
                    "platz": {
                        "type": "integer",
                        "example": 12
                    },
                    "untertitel": {
                        "type": "boolean",
                        "example": true
                    },
                    "snackzuschlag_chf": {
                        "type": "number",
                        "format": "float",
                        "example": 3.5,
                        "nullable": true
                    },
                    "film_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "typ_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "sprache_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "saal_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}